Refactor testing
