+++ .eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": "./node_modules/gts/"
+  "extends": "./node_modules/gts/",
+  "rules": {
+    "@typescript-eslint/no-namespace": "off"
+  }
 }
