+++ test/test_files/golden/strict_null_checks_sample.ts
